chore: update Makefile to support creating a signed helm package
Use "make helm-package HELM_KEY=" to avoid signing the package, or
"make helm-package HELM_KEY=your-key" to select the secret key from
your keyring to use for signing.

Signed packages are placed in a zip file along with the .prov file
which should be copied to whereever the helm package .tgz file is
published.