Host hostgroup support (#117)
feat: update api-go for Host and Host Group support in LUN selection

Update the seagate-exos-x-api-go library to 2.4.1 which supports
grouping of initiators into Hosts and Host Groups and fixes a bug in
volume creation on linear storage system. Also update grpc lib
for security updates. Update go version to 1.22 and driver version to 1.10